Se agregó autorización al sistema se manejan los siguientes roles: usuario, empleado, lider

+++ src/controllers/paquete.controller.js
@@ -18,6 +18,7 @@ const solicitarPaquete = async (req, res) => {
             direccionOrigen: direccionOrigen,
             ciudadDestino: ciudadDestino,
             direccionDestino: direccionDestino,
+            usuario: req.user._id,
         })
 
         await solicitudPaquete.save()
@@ -37,11 +38,37 @@ const solicitarPaquete = async (req, res) => {
 const eliminarSolicitudPaquete = async (req, res) => {
     try {
 
+        const { _id, tipoUsuario } = req.user
         const { id } = req.params;
 
-        const solicitudRemovida = await ModeloPauqete.findByIdAndRemove(id);
+        const solicitud = await ModeloPauqete.findById(id);
 
-        return res.status(200).json({ status: true, solicitud: solicitudRemovida })
+        if (!solicitud) {
+            return res.status(200).json({ status: true, message: "La solicitud no existe" })
+        }
+
+
+        if (tipoUsuario == "usuario") {
+
+            const solicitud = await ModeloPauqete.find({ _id: id, usuario: _id });
+
+            if (solicitud.length == 0) {
+                return res.status(403).json({ status: false, message: "El usuario no puede eliminar esta solicitud de paquete" })
+            }
+
+
+            const solicitudRemovida = await ModeloPauqete.findByIdAndRemove(id);
+
+            return res.status(200).json({ status: true, solicitud: solicitudRemovida })
+
+        } else if (tipoUsuario == "lider") {
+
+            const solicitudRemovida = await ModeloPauqete.findByIdAndRemove(id);
+
+            return res.status(200).json({ status: true, solicitud: solicitudRemovida })
+        }
+
+
 
     } catch (error) {
         console.log(error)
@@ -53,17 +80,44 @@ const eliminarSolicitudPaquete = async (req, res) => {
 
 const listarSolicitudesPaquetes = async (req, res) => {
     try {
+        const { _id, tipoUsuario } = req.user
 
-        const solicitudes = await ModeloPauqete.find();
+        if (tipoUsuario == "usuario") {
+            const solicitudes = await ModeloPauqete.find({ usuario: _id });
 
-        if (!solicitudes) {
-            return res.status(200).json({
-                status: true,
-                solicitudes: false
-            })
+            if (!solicitudes) {
+                return res.status(200).json({
+                    status: true,
+                    solicitudes: false
+                })
+            }
+
+            return res.status(200).json({ status: true, solicitudes: solicitudes })
+
+        } else if (tipoUsuario == "empleado") {
+            const solicitudes = await ModeloPauqete.find({ empleado: _id });
+
+            if (!solicitudes) {
+                return res.status(200).json({
+                    status: true,
+                    solicitudes: false
+                })
+            }
+
+            return res.status(200).json({ status: true, solicitudes: solicitudes })
+
+        } else if (tipoUsuario == "lider") {
+            const solicitudes = await ModeloPauqete.find();
+
+            if (!solicitudes) {
+                return res.status(200).json({
+                    status: true,
+                    solicitudes: false
+                })
+            }
+
+            return res.status(200).json({ status: true, solicitudes: solicitudes })
         }
-
-        return res.status(200).json({ status: true, solicitudes: solicitudes })
 
     } catch (error) {
         console.log(error)
@@ -76,16 +130,45 @@ const listarSolicitudesPaquetes = async (req, res) => {
 const listarSolicitudesPaquetesRecogidos = async (req, res) => {
     try {
 
-        const solicitudes = await ModeloPauqete.find({ "estado": true });
 
-        if (!solicitudes) {
-            return res.status(200).json({
-                status: true,
-                solicitudes: false
-            })
+        const { _id, tipoUsuario } = req.user
+
+        if (tipoUsuario == "usuario") {
+            const solicitudes = await ModeloPauqete.find({ usuario: _id, "estado": true });
+
+            if (!solicitudes) {
+                return res.status(200).json({
+                    status: true,
+                    solicitudes: false
+                })
+            }
+
+            return res.status(200).json({ status: true, solicitudes: solicitudes })
+
+        } else if (tipoUsuario == "empleado") {
+            const solicitudes = await ModeloPauqete.find({ empleado: _id, "estado": true });
+
+            if (!solicitudes) {
+                return res.status(200).json({
+                    status: true,
+                    solicitudes: false
+                })
+            }
+
+            return res.status(200).json({ status: true, solicitudes: solicitudes })
+
+        } else if (tipoUsuario == "lider") {
+            const solicitudes = await ModeloPauqete.find({ lider: _id, "estado": true });
+
+            if (!solicitudes) {
+                return res.status(200).json({
+                    status: true,
+                    solicitudes: false
+                })
+            }
+
+            return res.status(200).json({ status: true, solicitudes: solicitudes })
         }
-
-        return res.status(200).json({ status: true, solicitudes: solicitudes })
 
     } catch (error) {
         console.log(error)
@@ -98,16 +181,45 @@ const listarSolicitudesPaquetesRecogidos = async (req, res) => {
 const listarSolicitudesPaquetesNoRecogidos = async (req, res) => {
     try {
 
-        const solicitudes = await ModeloPauqete.find({ "estado": false });
 
-        if (!solicitudes) {
-            return res.status(200).json({
-                status: true,
-                solicitudes: false
-            })
+        const { _id, tipoUsuario } = req.user
+
+        if (tipoUsuario == "usuario") {
+            const solicitudes = await ModeloPauqete.find({ usuario: _id, "estado": false });
+
+            if (!solicitudes) {
+                return res.status(200).json({
+                    status: true,
+                    solicitudes: false
+                })
+            }
+
+            return res.status(200).json({ status: true, solicitudes: solicitudes })
+
+        } else if (tipoUsuario == "empleado") {
+            const solicitudes = await ModeloPauqete.find({ empleado: _id, "estado": false });
+
+            if (!solicitudes) {
+                return res.status(200).json({
+                    status: true,
+                    solicitudes: false
+                })
+            }
+
+            return res.status(200).json({ status: true, solicitudes: solicitudes })
+
+        } else if (tipoUsuario == "lider") {
+            const solicitudes = await ModeloPauqete.find({ lider: _id, "estado": false });
+
+            if (!solicitudes) {
+                return res.status(200).json({
+                    status: true,
+                    solicitudes: false
+                })
+            }
+
+            return res.status(200).json({ status: true, solicitudes: solicitudes })
         }
-
-        return res.status(200).json({ status: true, solicitudes: solicitudes })
 
     } catch (error) {
         console.log(error)
@@ -121,8 +233,9 @@ const listarSolicitudesPaquetesNoRecogidos = async (req, res) => {
 const editarEstadoSolicitudePaquete = async (req, res) => {
     try {
 
-        const { id } = req.params;
+        const { id } = req.params
         const { estado } = req.body
+        const { _id } = req.user
 
         if (estado != true && estado != false) {
             return res.status(400).json({
@@ -131,7 +244,7 @@ const editarEstadoSolicitudePaquete = async (req, res) => {
             })
         }
 
-        const solicitud = await ModeloPauqete.findByIdAndUpdate(id, { "estado": estado }, { new: true });
+        const solicitud = await ModeloPauqete.findByIdAndUpdate(id, { "estado": estado, "empleado": _id }, { new: true });
 
         if (!solicitud) {
             return res.status(200).json({
